salary-estimator/requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
